refactor(listener): extract memory list command
